Add entry about monthly teaching meet-up
As Lydia noted on Slack, this never made this page somehow (or was
accidentally removed?)  I've made it link to the community calendar
since we don't have a different/better landing page for it.

While here, I've emboldened the community calendar link to make it
stand out better, and added day/time information for the standing
live meetings.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>